requirements.txt
asyncio
asyncssh
PyYAML
